commit no.16

+++ project-ware/src/main/java/com/wwd/modules/ware/dao/WareSkuDao.java
@@ -13,6 +13,6 @@ import org.apache.ibatis.annotations.Mapper;
 @Mapper
 public interface WareSkuDao extends BaseDao<WareSkuEntity> {
 
-    Integer getSkuHasStockVoBySkuId(Long sku_id);
+    Long getSkuHasStockVoBySkuId(Long sku_id);
 	
 }
